requirements.txt
django>=4.2
pymupdf>=1.24.0
Pillow>=10.0
